Messed around with ForecastController, Added extra Mapper, stuck with forecastClientProperties not beeing initialized
