+++ src/main/java/com/parachute/booking/forecast/ForecastController.java
@@ -1,0 +1,33 @@
+package com.parachute.booking.forecast;
+
+import com.parachute.booking.forecast.api.ForecastClient;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/forecast")
+public class ForecastController {
+
+    private final DayCreator dayCreator;
+    private final ForecastClient forecastClient;
+
+    @GetMapping
+    @ResponseStatus(HttpStatus.OK)
+    public void getForecasts(){
+
+    }
+
+    @GetMapping("/{year}-{month}-{day}")
+    @ResponseStatus(HttpStatus.OK)
+    public List<ForecastDto> getListOfForecastsForDay(@PathVariable int year,@PathVariable int month,@PathVariable int day){
+
+        return forecastClient.getForecast(dayCreator.
+                createCalendarDayFromNumbers(year, month, day));
+    }
+
+
+}
